agentic_work/19_advance_tools/pyproject.toml
[project]
name = "19-advance-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.8",
]
